Rename payment module to fee
