+++ src/main/java/com/example/demox/application/ParkingMeterService.java
@@ -1,7 +1,7 @@
 package com.example.demox.application;
 
 import com.example.demox.domain.model.driver.DriverId;
-import com.example.demox.domain.model.payment.Fee;
+import com.example.demox.domain.model.fee.Fee;
 import com.example.demox.domain.model.ticket.NumberPlate;
 import com.example.demox.domain.model.ticket.Ticket;
 import com.example.demox.domain.model.ticket.TicketId;
