+++ src/main/java/com/example/demox/domain/model/fee/FeeRepository.java
@@ -1,0 +1,9 @@
+package com.example.demox.domain.model.fee;
+
+import java.util.Collection;
+import java.util.Date;
+
+public interface FeeRepository  {
+    void store(Fee fee);
+    Collection<Fee> getFeesByDayOfYear(Date date);
+}
