--- src/main/java/com/example/demox/domain/model/payment/FeeRepository.java
@@ -1,9 +1,0 @@
-package com.example.demox.domain.model.payment;
-
-import java.util.Collection;
-import java.util.Date;
-
-public interface FeeRepository  {
-    void store(Fee fee);
-    Collection<Fee> getFeesByDayOfYear(Date date);
-}
